# UTF8 processing: performance improvements, validation, and other useful APIs