--- a/src/System.Text.Primitives/System.Text.Primitives.csproj
+++ b/src/System.Text.Primitives/System.Text.Primitives.csproj
@@ -11,4 +11,7 @@
   <ItemGroup>
     <ProjectReference Include="..\System.Buffers.Primitives\System.Buffers.Primitives.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="System.Numerics.Vectors" Version="$(CoreFxStableVersion)" />
+  </ItemGroup>
 </Project>