… src/System.Text.Primitives/System/Text/Encoders/Utf8Util.Helpers.cs
… src/System.Text.Primitives/System/Text/Encoders/Utf8Util.Validation.cs
… src/System.Text.Primitives/System/Text/Encoders/Utf8Utility.cs
… tests/System.Text.Primitives.Tests/Encoding/Utf8UtilTests.Validation.cs
… tests/System.Text.Primitives.Tests/Encoding/Utf8UtilityTests.cs
… tests/System.Text.Primitives.Tests/MemoryProtection/NativeMemory.Windows.cs
… tests/System.Text.Primitives.Tests/SampleTexts/11-0.txt
… tests/System.Text.Primitives.Tests/SampleTexts/11.txt
… tests/System.Text.Primitives.Tests/SampleTexts/25249-0.txt
… tests/System.Text.Primitives.Tests/SampleTexts/30774-0.txt
… tests/System.Text.Primitives.Tests/SampleTexts/39251-0.txt