This PR introduces a workhorse routine for validating UTF-8 strings without performing a proper to-UTF16 conversion. It is intended to be used as the backend to UTF-8 validation routines, including routines that perform substitution of invalid sequences to convert an ill-formed UTF-8 string into a well-formed UTF-8 string. There are friendly APIs on the `Utf8Utility` type that allow inspecting the UTF-8 string. These APIs are intended to be compatible with `System.Text.Encoding.UTF8` in that they agree on the length of bad sequences, etc.

There is a sister workhorse "convert UTF-8 to UTF-16" routine, but it's not part of this PR since I'm still adding unit tests for it.

The unit tests for the workhorse routine perform some limited self-fuzzing. Additionally, the unit test project introduces the `NativeMemory` type (for unit test use only), which grabs memory pages directly from the kernel's page allocator and bookends the returned range with poison pages, where accessing these pages will immediately AV the process. This type helps detect buffer overruns in code that works directly with memory buffers in an unsafe fashion.

Finally, this API introduces substantial performance improvements over the existing "how many UTF-16 / UTF-32 code units would there be if I were to convert this UTF-8 string to one of those representations?" framework API `System.Text.Encoding.UTF8Encoding.GetChars`. I've copied the performance improvement table below.

# Perf results

1,000,000 iterations on each lipsum

Testbed: Intel(R) Core(TM) i7-6700 CPU @ 3.40GHz, 3408 Mhz, 8 Core(s), 8 Logical Processor(s), Win10 RS3 amd64fre

## English (ASCII)

* `System.Text.Encoding.UTF8`: 0.286 sec
* `Utf8Util.GetCharCount`: 0.077 sec

73% reduction in runtime

## Hebrew (primarily 2-byte)

* `System.Text.Encoding.UTF8`: 5.22 sec
* `Utf8Util.GetCharCount`: 1.81 sec

65% reduction in runtime

## Cyrillic (primarily 2-byte)

* `System.Text.Encoding.UTF8`: 5.20 sec
* `Utf8Util.GetCharCount`: 1.78 sec

65% reduction in runtime

## Japanese (primarily 3-byte)

* `System.Text.Encoding.UTF8`: 3.06 sec
* `Utf8Util.GetCharCount`: 1.08 sec

65% reduction in runtime

## Chinese (primarily 3-byte)

* `System.Text.Encoding.UTF8`: 6.30 sec
* `Utf8Util.GetCharCount`: 2.25 sec

65% reduction in runtime

The corpus that was used for testing can be found at https://github.com/GrabYourPitchforks/fast-utf8/blob/master/FastUtf8Tester/Lipsum.cs. That same directory contains the test program and full results.