@ahsonkhan 2017-10-26
>  I've copied the performance improvement table below.

Can you please check-in the performance tests to this repo (preferably as part of this PR) so that we can run them using xunit.performance and our perf harness?