@GrabYourPitchforks 2017-10-26
@ahsonkhan I just added the perf tests and sample texts. These are different than the sample texts I had been using, so the numbers will be somewhat off.