@Drawaes 2017-10-26
@GrabYourPitchforks  +++10 I would love this to become the start of more fuzzing ability for .net as it's something I have been wanting to get around to for a while...