@GrabYourPitchforks 2017-11-08
Since the bulk of the logic isn't changing and we're discussing utility methods / renamings, I'm going to merge what I have now so that people can immediately start playing with it. Will send a new PR with additional performance improvements and changes.